Remove iconv-lite from dependencies and fix grunt task for deployment
Looks like you don't need to convert the buffer which is read from the zip file before uploading it.
This means iconv-lite is also not necessary anymore.
The deployment task tried to deploy the app directly without building it first. Now it will build it first and then deploy it.